+++ lib/src/formatters/format_match.dart
@@ -1,0 +1,16 @@
+import 'package:moment_dart/src/formatters/token.dart';
+
+class FormatMatch {
+  final int startIndex;
+  final FormatterToken token;
+
+  int get endIndex => token.name.length + startIndex;
+
+  const FormatMatch({
+    required this.startIndex,
+    required this.token,
+  });
+
+  @override
+  String toString() => token.name;
+}