+++ lib/src/localizations/mn_TM.dart
@@ -87,7 +87,7 @@ class LocalizationMongolianTraditional extends Localization {
   String weekdayName(int i) => weekdayNames[i]!;
 
   @override
-  String calendar(Moment moment, {Moment? reference, bool weekStartOnSunday = false, bool omitHours = false}) {
+  String calendar(Moment moment, {Moment? reference, bool weekStartOnSunday = false, bool omitHours = false, String? customFormat}) {
     reference ??= Moment.now();
 
     late final String day;
@@ -112,7 +112,7 @@ class LocalizationMongolianTraditional extends Localization {
         final Moment startOfLastWeek = Localization.weekFirstDay(reference).subtract(const Duration(days: 7));
 
         if (moment.isBefore(startOfLastWeek)) {
-          day = moment.format(localizationDefaultDateFormat());
+          day = moment.format(customFormat ?? localizationDefaultDateFormat());
         } else {
           day = "ᠡᠮᠦᠨᠡᠬᠢ ${weekdayName(moment.dateTime.weekday)}";
         }
@@ -132,9 +132,13 @@ class LocalizationMongolianTraditional extends Localization {
         if (moment.isBefore(startOfNextWeek)) {
           day = weekdayName(moment.dateTime.weekday);
         } else {
-          day = moment.format(localizationDefaultDateFormat());
+          day = moment.format(customFormat ?? localizationDefaultDateFormat());
         }
       }
+    }
+
+    if (customFormat != null) {
+      return day;
     }
 
     if (omitHours) {