+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2
+
+- Improved `Moment.format` tokenization
+- `Moment.calender` now allows `customFormat`, which can be used instead of the localization's default format.
+- Fixes
+
 ## 0.1.1
 
 - Fixed `Moment.format` function